Tweak

+++ StateMachine.cpp
@@ -9,7 +9,7 @@ void StateMachine::setState(int state) {
     this->state = state;
 }
 
-void StateMachine::process(uint8_t value) {
+void StateMachine::process(int value) {
   for (int i = 0; this->jumpVector[i].handler != 0; i++) {
     if (this->jumpVector[i].state == this->state) {
       this->state = (this->jumpVector[i].handler)(value);
